hotfix

+++ examples/pytorch/gcn/gcn_mp.py
@@ -137,9 +137,9 @@ def main(args):
       #Val samples %d
       #Test samples %d""" %
           (n_edges, n_classes,
-              train_mask.sum().item(),
-              val_mask.sum().item(),
-              test_mask.sum().item()))
+              train_mask.int().sum().item(),
+              val_mask.int().sum().item(),
+              test_mask.int().sum().item()))
 
     if args.gpu < 0:
         cuda = False
